v0.5.0a: fixed alignment inconsistency in renderedtext

+++ SPD-classes/src/main/java/com/watabou/gltextures/SmartTexture.java
@@ -119,11 +119,11 @@ public class SmartTexture extends Texture {
 		bitmap = null;
 	}
 	
-	public RectF uvRect( int left, int top, int right, int bottom ) {
+	public RectF uvRect( float left, float top, float right, float bottom ) {
 		return new RectF(
-			(float)left		/ width,
-			(float)top		/ height,
-			(float)right	/ width,
-			(float)bottom	/ height );
+			left		/ width,
+			top		/ height,
+			right	/ width,
+			bottom	/ height );
 	}
 }
